requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
djangorestframework==3.12.4
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.2
Pillow==8.3.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rsa==4.8
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.6
